make perf tests easier to run with pnpm test:perf
